Making module loading for fglrx and nvidia not fail.
It has been observed that on some weird systems,
the graphic card is supported by fglrx or nvidia
(by having its pci-id listed in the drivers own
pci-id list), yet the driver fails to initialize
the card (shown by a 'device is busy' message).

In such cases, we need to not error-out of live-config
and continue as normal, then, xorg takes care about this
automatically loads the next best suitable driver.