Fix hook for disabling display managers
Init scripts of current display manager packages explicitly check the
existence of /etx/X11/default-display-manager and *do* continue the start
procedure if it doesn't exist (checked with kdm, lxdm, sddm, wdm, xdm). So
instead of deleting the file, set its content to something that just won't
match the respective string check for the DM name in the init script.